requirements.txt
asgiref==3.8.1
asttokens==2.4.1
colorama==0.4.6
comm==0.2.2
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.11
django-environ==0.11.2
django-svg-image-form-field==1.0.1
environ==1.0
exceptiongroup==1.2.2
executing==2.0.1
importlib_metadata==8.0.0
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.1
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
llvmlite==0.43.0
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numba==0.60.0
numpy==2.0.0
packaging==24.1
paho-mqtt==2.1.0
pandas==2.2.2
parso==0.8.4
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
psycopg2==2.9.9
pure-eval==0.2.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
pyts==0.13.0
pytz==2024.1
pywin32==306
pyzmq==26.0.3
scikit-learn==1.5.1
scipy==1.13.1
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
threadpoolctl==3.5.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
wcwidth==0.2.13
zipp==3.19.2
